requirements.txt
aiokafka==0.13.0
asyncpg==0.31.0
fastapi==0.133.0
httpx==0.28.1
mlflow==3.8.1
mlflow_skinny==3.8.1
mlflow_tracing==3.8.1
numpy==2.4.2
pydantic==2.12.5
pytest==9.0.2
pytest_asyncio==1.3.0
python-dotenv==1.2.1
redis==7.2.0
Requests==2.32.5
scikit_learn==1.8.0
starlette==0.52.1
uvicorn==0.41.0
